+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/ColumnConditionPanel.java
@@ -121,9 +121,9 @@ public abstract class ColumnConditionPanel<C, T> extends JPanel {
 
 	/**
 	 * The default implementation returns an empty Optional.
-	 * @return an event notified when a subcomponent of this condition panel receives focus or an empty Optional if none is available
+	 * @return an observer notified when a subcomponent of this condition panel receives focus or an empty Optional if none is available
 	 */
-	public Optional<EventObserver<C>> focusGainedEvent() {
+	public Optional<EventObserver<C>> focusGainedObserver() {
 		return Optional.empty();
 	}