+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 - FilterTable, now propagates the associated MouseEvent as the source of the ActionEvent when doubleClickAction is performed.
 - MenuBuilder no longer extends MenuItemBuilder.
 - ComponentDialogBuilder.closeEvent() renamed closeObserver().
+- ColumnConditionPanel.focusGainedEvent() renamed focusGainedObserver().
 ### is.codion.swing.framework.ui
 - EntityTablePanel bug fixed, Config.SHOW_REFRESH_PROGRESS_BAR now used for default config value.
 - EntityDialogs bug fixed, now respects the location if specified.